requirements.txt
anyio==3.7.0
certifi==2023.5.7
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
